requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
python-dotenv==1.0.0
watchdog==3.0.0 